fix(cli): move files
